✅ Task 7完了: TestSupportTools.js分割 - Main Controller Pattern適用
## 主要な変更内容

### TestSupportTools.js分割（2,527語→497語、80%削減）
- Main Controller Patternによる軽量オーケストレーター化
- 4つの専門コンポーネントに機能分離
- 全公開API互換性維持、既存依存関係保持

### 新規作成コンポーネント（4件）
1. **TestExecutionManager.js** - テスト実行・スイート調整
   - TestRunner統合、デフォルトテストスイート管理
   - 個別テストメソッド（GameEngine、BubbleManager、ScoreManager等）
   - パフォーマンステスト（レンダリング、メモリ、FPS安定性）

2. **MockDataManager.js** - モックデータライフサイクル・キャッシュ管理
   - MockDataGenerator統合、キャッシュシステム実装
   - バブル・ゲーム状態・プレイヤーデータ生成
   - メモリ効率的なデータクローン・キャッシュ制限

3. **BenchmarkManager.js** - ベンチマーク実行・スイート管理
   - BenchmarkSuite統合、履歴管理
   - 5種類ベンチマーク（レンダリング、物理演算、パーティクル、メモリ、音声）
   - ボトルネック特定・改善提案システム

4. **TestResultProcessor.js** - テスト結果処理・フォーマット
   - 結果分析・トレンド分析・推奨事項生成
   - CSV/JSON/TXT形式エクスポート機能
   - 健全性評価・レポート生成

### 技術仕様
- **BaseComponent継承**: 統一エラーハンドリング・初期化パターン
- **依存注入**: Main Controllerによるコンポーネント管理
- **後方互換性**: 旧APIアクセス（testRunner、mockDataGenerator等）
- **エラー処理**: 包括的try-catch、グレースフル劣化対応

### プロジェクト進捗
- **Phase F.2進捗**: Task 1-7完了（10タスク中70%）
- **MCPツール互換性**: 2,500語以下ファイル分割要件達成
- **残作業**: GameStateCommands分割（Task 8）、統合テスト（Task 9-10）
