+++ .kiro/specs/debug-test-file-splitting-phase-f2/tasks.md
@@ -191,38 +191,38 @@
   - Test suggestion generation and history management features
   - _Requirements: 10.1, 10.3, 11.1_
 
-- [ ] 7. Split TestSupportTools.js (2,527 words)
-- [ ] 7.1 Create TestExecutionManager component
+- [x] 7. Split TestSupportTools.js (2,527 words)
+- [x] 7.1 Create TestExecutionManager component
   - Extract test execution and suite coordination logic
   - Implement execution monitoring and state management
   - Create unit tests for test execution functionality
   - _Requirements: 7.2, 7.5_
 
-- [ ] 7.2 Create MockDataManager component
+- [x] 7.2 Create MockDataManager component
   - Extract mock data lifecycle and caching logic
   - Implement data cleanup and management operations
   - Create unit tests for mock data management functionality
   - _Requirements: 7.2, 7.5_
 
-- [ ] 7.3 Create BenchmarkManager component
+- [x] 7.3 Create BenchmarkManager component
   - Extract benchmark operations and suite management logic
   - Implement performance measurement and reporting
   - Create unit tests for benchmark management functionality
   - _Requirements: 7.2, 7.5_
 
-- [ ] 7.4 Create TestResultProcessor component
+- [x] 7.4 Create TestResultProcessor component
   - Extract test result processing and formatting logic
   - Implement result analysis and reporting
   - Create unit tests for test result processing functionality
   - _Requirements: 7.2, 7.5_
 
-- [ ] 7.5 Refactor TestSupportTools main controller
+- [x] 7.5 Refactor TestSupportTools main controller
   - Update main controller to coordinate test support components
   - Maintain test environment and execution state management
   - Implement component initialization and error handling
   - _Requirements: 7.1, 7.3, 7.4_
 
-- [ ] 7.6 Test TestSupportTools integration
+- [x] 7.6 Test TestSupportTools integration
   - Run existing tests to ensure test support functionality is preserved
   - Verify test execution and mock data management work correctly
   - Test benchmark operations and result processing features
